Inherit ClientResponse from HeadersMixin. Change `content_type` property behavior if no Content-Type header.
