+++ aiohttp/client_reqrep.py
@@ -14,7 +14,7 @@ from yarl import URL
 import aiohttp
 
 from . import hdrs, helpers, streams
-from .helpers import Timeout
+from .helpers import HeadersMixin, Timeout
 from .log import client_logger
 from .multipart import MultipartWriter
 from .protocol import HttpMessage
@@ -488,7 +488,7 @@ class ClientRequest:
             self._writer = None
 
 
-class ClientResponse:
+class ClientResponse(HeadersMixin):
 
     # from the Status-Line of the response
     version = None  # HTTP-Version
@@ -582,28 +582,6 @@ class ClientResponse:
     def history(self):
         """A sequence of of responses, if redirects occurred."""
         return self._history
-
-    @property
-    def content_type(self):
-        """The value of content part for Content-Type HTTP header."""
-        ctype = self.headers.get(hdrs.CONTENT_TYPE, None)
-        if ctype is None:
-            # no Content-Type header
-            return None
-        mtype, stype, _, _ = helpers.parse_mimetype(ctype.lower())
-        _content_type = '{mtype}/{stype}'.format(mtype=mtype, stype=stype)
-        return _content_type
-
-    @property
-    def charset(self):
-        """The value of charset part for Content-Type HTTP header."""
-        ctype = self.headers.get(hdrs.CONTENT_TYPE, None)
-        if ctype is None:
-            # no Content-Type header
-            return None
-        _, _, _, params = helpers.parse_mimetype(ctype.lower())
-        _charset = params.get('charset')
-        return _charset
 
     def waiting_for_continue(self):
         return self._continue is not None
