+++ aiohttp/helpers.py
@@ -3,6 +3,7 @@
 import asyncio
 import base64
 import binascii
+import cgi
 import datetime
 import functools
 import io
@@ -620,3 +621,44 @@ class TimeService:
         if s is None:
             self._strtime = s = self._format_date_time()
         return self._strtime
+
+
+class HeadersMixin:
+
+    _content_type = None
+    _content_dict = None
+    _stored_content_type = sentinel
+
+    def _parse_content_type(self, raw):
+        self._stored_content_type = raw
+        if raw is None:
+            # default value according to RFC 2616
+            self._content_type = 'application/octet-stream'
+            self._content_dict = {}
+        else:
+            self._content_type, self._content_dict = cgi.parse_header(raw)
+
+    @property
+    def content_type(self, _CONTENT_TYPE=hdrs.CONTENT_TYPE):
+        """The value of content part for Content-Type HTTP header."""
+        raw = self.headers.get(_CONTENT_TYPE)
+        if self._stored_content_type != raw:
+            self._parse_content_type(raw)
+        return self._content_type
+
+    @property
+    def charset(self, _CONTENT_TYPE=hdrs.CONTENT_TYPE):
+        """The value of charset part for Content-Type HTTP header."""
+        raw = self.headers.get(_CONTENT_TYPE)
+        if self._stored_content_type != raw:
+            self._parse_content_type(raw)
+        return self._content_dict.get('charset')
+
+    @property
+    def content_length(self, _CONTENT_LENGTH=hdrs.CONTENT_LENGTH):
+        """The value of Content-Length HTTP header."""
+        l = self.headers.get(_CONTENT_LENGTH)
+        if l is None:
+            return None
+        else:
+            return int(l)
