+++ tests/test_client_response.py
@@ -329,7 +329,7 @@ def test_content_type_no_header():
     response = ClientResponse('get', URL('http://def-cl-resp.org'))
     response.headers = {}
 
-    assert response.content_type is None
+    assert 'application/octet-stream' == response.content_type
 
 
 def test_charset():
